Update Bus Reservation Entity 내용 : 패키지 수정 및 예외 처리 #39
