+++ Everyone-Tayo-Server/src/main/java/team/sw/everyonetayo/api/busroute/BusRouteController.java
@@ -1,4 +1,4 @@
-package team.sw.everyonetayo.util.busroute;
+package team.sw.everyonetayo.api.busroute;
 
 import org.json.JSONArray;
 import org.json.XML;
@@ -14,7 +14,6 @@ import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
