+++ Everyone-Tayo-Server/src/main/java/team/sw/everyonetayo/api/busroute/Route.java
@@ -1,4 +1,4 @@
-package team.sw.everyonetayo.util.busuuid;
+package team.sw.everyonetayo.api.busroute;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -16,12 +16,19 @@ import javax.persistence.Id;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class BusUuid {
+public class Route {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String uuid; // 노선 유형
-    private String routenm; //노선 번호
+    private String routeTp; // 노선 유형
+    private String routeId; //노선 id
+//    private String startVehicleTime;  // 첫차 시간
+//    private String endVehicleTime; // 막차 시간
+//    private String startNodeNm; // 기점
+//    private String endNodeNm; // 종점
+    private String routeNo; // 노선 번호
     private String cityCode; // 도시 코드
+
 }
