+++ Everyone-Tayo-Server/src/main/java/team/sw/everyonetayo/api/busstop/BusStop.java
@@ -1,4 +1,4 @@
-package team.sw.everyonetayo.util.busstop;
+package team.sw.everyonetayo.api.busstop;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
