+++ Everyone-Tayo-Server/src/main/java/team/sw/everyonetayo/api/busstop/BusStopController.java
@@ -1,11 +1,10 @@
-package team.sw.everyonetayo.util.busstop;
+package team.sw.everyonetayo.api.busstop;
 
 import org.json.JSONArray;
 import org.json.XML;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
