+++ Everyone-Tayo-Server/src/main/java/team/sw/everyonetayo/api/busstop/BusStopRepository.java
@@ -1,10 +1,9 @@
-package team.sw.everyonetayo.util.busstop;
+package team.sw.everyonetayo.api.busstop;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface BusStopRepository extends JpaRepository<BusStop, Long> {
