+++ Everyone-Tayo-Server/src/main/java/team/sw/everyonetayo/api/busstop/BusStopService.java
@@ -1,10 +1,9 @@
-package team.sw.everyonetayo.util.busstop;
+package team.sw.everyonetayo.api.busstop;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class BusStopService {
