+++ Everyone-Tayo-Server/src/main/java/team/sw/everyonetayo/api/busuuid/BusUuidDto.java
@@ -1,4 +1,4 @@
-package team.sw.everyonetayo.util.busuuid;
+package team.sw.everyonetayo.api.busuuid;
 
 import lombok.Builder;
 import lombok.Getter;
