+++ Everyone-Tayo-Server/src/main/java/team/sw/everyonetayo/reservation/ResponseNearBusDto.java
@@ -1,4 +1,4 @@
-package team.sw.everyonetayo.util.reservation;
+package team.sw.everyonetayo.reservation;
 
 import lombok.Builder;
 import lombok.Data;
